requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14.0
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.2.0 